Some refactoring and file reorder
